// server/enums/sessionSortBy.ts
/**
 * Enum of session search page 'sort by' columns. Values aligned with the UI.
 */
enum SessionSortBy {
  NAME = 'name',
  RELEASE_DATE = 'release-date',
  LOCATION = 'location',
  SESSION_TYPE = 'session-type',
  DUE_BY = 'due-by',
  EXEMPTION_DATE = 'exemption-entered-on',
  EXEMPTION_REASON = 'exemption-reason',
}

export default SessionSortBy