// server/enums/sessionSearchSortField.ts
enum SessionSearchSortField {
  PRISONER_NAME = 'PRISONER_NAME',
  PRISON_NUMBER = 'PRISON_NUMBER',
  CELL_LOCATION = 'CELL_LOCATION',
  RELEASE_DATE = 'RELEASE_DATE',
  SESSION_TYPE = 'SESSION_TYPE',
  DUE_BY = 'DUE_BY',
  EXEMPTION_DATE = 'EXEMPTION_DATE',
  EXEMPTION_REASON = 'EXEMPTION_REASON',
}

export default SessionSearchSortField